+++ src/core/common/Sidebar/sortResourcesArray.js
@@ -1,12 +1,7 @@
 import { get } from "lodash";
 import { resourceOrder } from "../../../version/config/theme.config";
 
-export const SHORT_MENU = "short";
-export const FULL_MENU = "full";
-
 /**
- * This function sorts resources by theme settings
- *
  * @author Bogdan Shcherban <bsc@piogroup.net>
  * @param {array} resources
  * @return {array}
@@ -23,24 +18,4 @@ export function sortResourcesArray(resources) {
         }
     }
     return sortResource;
-}
-
-/**
- * This function returns type of menu
- *
- * @author Bogdan Shcherban <bsc@piogroup.net>
- * @param {string} currentPathname
- * @return {string}
- */
-export function getMenuType(currentPathname) {
-    let result = SHORT_MENU;
-    const ShortMenuPages = [
-        "/",
-        "/charts",
-        "/patients"
-    ];
-    if (ShortMenuPages.indexOf(currentPathname) === -1) {
-        result = FULL_MENU;
-    }
-    return result;
 }
