+++ src/core/common/Topbar/functions.js
@@ -1,11 +1,9 @@
 /**
- * This function checks is current page should have a title
- *
  * @author Bogdan Shcherban <bsc@piogroup.net>
  * @param {shape} location
  * @return {boolean}
  */
-export function isPageHasTitle(location) {
+export function pageHasTitle(location) {
     const pathName = location.pathname;
     const pagesWithTitle = [
         '/',
